Struct duckscript::types::command::Commands

pub struct Commands {
    pub commands: HashMap<String, CommandBox>,
    pub aliases: HashMap<String, String>,
}

Holds and enables access to the runtime commands implementations

Fields
commands: HashMap<String, CommandBox>

mapping between command names to implementations

aliases: HashMap<String, String>

mapping between aliases to command names

Methods

impl Commands

pub fn new() -> Commands

Creates and returns a new instance.

pub fn set(&mut self, command: CommandBox) -> Result<(), ScriptError>

Adds a new command definition. It will fail in case another command already defined the same name/aliases

pub fn get(&self, name: &str) -> Option<&CommandBox>

Return the command based on the given command name/alias

pub fn exists(&self, name: &str) -> bool

Return true if the command based on the given command name/alias exists

pub fn get_for_use(&mut self, name: &str) -> Option<CommandBox>

Return the command based on the given command name/alias. It will also remove it in the process.

pub fn get_all_command_names(&self) -> Vec<String>

Returns all the command names currently registered

pub fn remove(&mut self, name: &str) -> bool

Removes the requested command.